recommendation_engine/ml-recommendation-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==2.1.1
pydantic==2.9.2
redis==5.1.1
faiss-cpu==1.9.0.post1
